requirements.txt
firecrawl-py
pandas
requests
python-dotenv
beautifulsoup4
markdown
